churn-prediction-system/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
mysql-connector-python
sqlalchemy
python-multipart
joblib
